chore(nav): retire the bottomNav feature flag
The bottom nav and desktop sidebar shipped in #184 and have been on in
production since. The flag was still off in dev, unit tests, and e2e, so
every non-production surface rendered the legacy top Header that no user
sees. Removing it makes those environments match production and lets the
Header go, along with the nav destination list it duplicated from
~/lib/navItems.
