+++ .env.production
@@ -1,3 +1,2 @@
-VITE_FEATURE_BOTTOMNAV=true
 VITE_FEATURE_EXPLORE=false
 VITE_FEATURE_WEEKLY_BALANCE=false
