+++ src/components/BottomNav/BottomNav.stories.tsx
@@ -9,8 +9,8 @@ import { BottomNav } from './BottomNav';
 
 // The bar reads flags from the session-aware `useFeatures()` hook and is
 // desktop-hidden (`sm:hidden`). Use an owner session + the preview override so
-// every feature (including `bottomNav`) resolves on, and view it in a mobile
-// viewport to see the bar.
+// every feature resolves on and the optional tabs appear, and view it in a
+// mobile viewport to see the bar.
 const ownerSession = {
   user: { email: 'daniel_bowers@icloud.com' },
 } as unknown as Session;
