+++ src/components/BottomNav/useBottomNavVisible.test.tsx
@@ -2,28 +2,8 @@ import { act, renderHook, screen } from '@testing-library/react';
 import { fireEvent, render } from '@testing-library/react';
 import { ReactNode } from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import { Mock, beforeEach, vi } from 'vitest';
-
-import { Features } from '~/config/features';
-import { useFeatures } from '~/hooks';
 
 import { useBottomNavVisible } from './useBottomNavVisible';
-
-vi.mock('~/hooks', () => ({
-  useFeatures: vi.fn(),
-}));
-
-const mockedUseFeatures = useFeatures as unknown as Mock;
-
-const featuresWith = (overrides: Partial<Features> = {}): Features => ({
-  bottomNav: true,
-  explore: false,
-  premium: false,
-  programs: false,
-  spotify: false,
-  weeklyBalance: false,
-  ...overrides,
-});
 
 const wrapperFor =
   (initialPath = '/') =>
@@ -31,29 +11,15 @@ const wrapperFor =
     <MemoryRouter initialEntries={[initialPath]}>{children}</MemoryRouter>
   );
 
-beforeEach(() => {
-  mockedUseFeatures.mockReset();
-});
-
 describe('useBottomNavVisible', () => {
-  test('is visible when the flag is on and the route is not suppressed', () => {
-    mockedUseFeatures.mockReturnValue(featuresWith());
+  test('is visible when the route is not suppressed', () => {
     const { result } = renderHook(() => useBottomNavVisible(), {
       wrapper: wrapperFor('/'),
     });
     expect(result.current).toBe(true);
   });
 
-  test('is hidden when the bottomNav flag is off', () => {
-    mockedUseFeatures.mockReturnValue(featuresWith({ bottomNav: false }));
-    const { result } = renderHook(() => useBottomNavVisible(), {
-      wrapper: wrapperFor('/'),
-    });
-    expect(result.current).toBe(false);
-  });
-
   test('is hidden on the immersive /active route', () => {
-    mockedUseFeatures.mockReturnValue(featuresWith());
     const { result } = renderHook(() => useBottomNavVisible(), {
       wrapper: wrapperFor('/active'),
     });
@@ -61,7 +27,6 @@ describe('useBottomNavVisible', () => {
   });
 
   test('is hidden while a text input is focused (mobile keyboard)', () => {
-    mockedUseFeatures.mockReturnValue(featuresWith());
     const Probe = () => (
       <span data-testid="state">{String(useBottomNavVisible())}</span>
     );
