--- src/components/Header.styles.css
@@ -1,3 +1,0 @@
-.active {
-  @apply bg-accent text-accent-foreground;
-}
