+++ src/pages/StartWorkoutPage/StartWorkoutPage.collapsibleSections.test.jsx
@@ -27,7 +27,6 @@ vi.mock('~/hooks', async (importOriginal) => ({
 }));
 
 const BASE_FEATURES = {
-  bottomNav: false,
   explore: false,
   premium: false,
   programs: true,
