+++ src/pages/StartWorkoutPage/StartWorkoutPage.hooksOrder.test.jsx
@@ -30,7 +30,6 @@ vi.mock('~/hooks', async (importOriginal) => ({
 }));
 
 const BASE_FEATURES = {
-  bottomNav: false,
   explore: false,
   premium: false,
   programs: true,
