+++ src/pages/StartWorkoutPage/StartWorkoutPage.parallelPrograms.test.jsx
@@ -37,7 +37,6 @@ vi.mock('~/hooks', async (importOriginal) => ({
 }));
 
 const BASE_FEATURES = {
-  bottomNav: false,
   explore: false,
   premium: false,
   programs: true,
