+++ src/pages/StartWorkoutPage/StartWorkoutPage.startProgramCard.test.jsx
@@ -23,7 +23,6 @@ vi.mock('~/hooks', async (importOriginal) => ({
 }));
 
 const BASE_FEATURES = {
-  bottomNav: false,
   explore: false,
   premium: false,
   programs: true,
