+++ vitest.config.ts
@@ -14,7 +14,6 @@ export default mergeConfig(
       // suite renders. Pin every build-time flag here (matching `.env.test`) so
       // runs are identical on every machine and in CI.
       env: {
-        VITE_FEATURE_BOTTOMNAV: 'false',
         VITE_FEATURE_EXPLORE: 'true',
         VITE_FEATURE_PREMIUM: 'false',
         VITE_FEATURE_PROGRAMS: 'true',
